Перенёс функцию для определения продолжительности события во вспомогательные функции
